requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
python-dotenv
rich
click
requests
